Added feature: log-in

+++ Assets/Scripts/Entities/DojoModels/States/MatchState/MatchPlayerCharacter.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Numerics;
 using Dojo;
 using Dojo.Torii;
 using dojo_bindings;
@@ -8,12 +9,12 @@ public class MatchPlayerCharacter : ModelInstance
 {
     private UInt32 match_id;
     private dojo.FieldElement player;
-    private int playerId;
+    private BigInteger playerId;
     private UInt32 id;
     private UInt32 character_id;
 
     public uint Match_id { get => match_id; set => match_id = value; }
-    public int PlayerId { get => playerId; set => playerId = value; }
+    public BigInteger PlayerId { get => playerId; set => playerId = value; }
     public uint Character_id { get => character_id; set => character_id = value; }
 
     public override void Initialize(Model model)
@@ -24,7 +25,7 @@ public class MatchPlayerCharacter : ModelInstance
         character_id = model.members["character_id"].ty.ty_primitive.u32;
 
         var playerString = BitConverter.ToString(player.data.ToArray()).Replace("-", "").ToLower();
-        playerId = System.Int32.Parse(playerString, System.Globalization.NumberStyles.AllowHexSpecifier);
+        playerId = BigInteger.Parse(playerString, System.Globalization.NumberStyles.AllowHexSpecifier);
 
         // Debug.Log("MatchPlayerCharacter: \n match_id: " + match_id
         //           + "\n playerId: " + playerId
